fix(b2b-flows): correct increment of flow version on re-deployment
Previously, the flow version was erroneously incremented by 3 instead of
the intended 1 upon re-deployment. This was due to the usage of the version
field as versionKey in the mongoose schema. Consequently, whenever a
document was saved using doc.save(), mongoose would inadvertently increase
the version number. This commit rectifies the issue by manually setting the
version increment.

Jira Issue: DS-421